Refacter modelDTO to Model in All Repository, done Change Student selected Status, Teacher, Staff, Class, Room, Subject, ExamSchedule, Status selected, Done to refacter All Controller from APIResponse to IActionResult

+++ USMS_BackendSolution/SchedulerService/Repository/StudentInExamScheduleRepository/StudentInExamScheduleRepository.cs
@@ -1,67 +1,31 @@
 ﻿using ISUZU_NEXT.Server.Core.Extentions;
 using Microsoft.EntityFrameworkCore;
 using SchedulerBusinessObject.AppDBContext;
-using SchedulerBusinessObject.ModelDTOs;
 using SchedulerBusinessObject.SchedulerModels;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 
 namespace SchedulerService.Repository.StudentInExamScheduleRepository
     {
     public class StudentInExamScheduleRepository : IStudentInExamScheduleRepository
         {
-
         #region Get All Student In Exam Schedule
         /// <summary>
         /// Get All Student In Exam Schedule Class
         /// </summary>
-        /// <returns></returns>
-        public async Task<List<StudentInExamScheduleDTO>> GetAllStudentInExamSchedule()
-            {
-            try
-                {
-                using (var _dbContext = new MyDbContext())
-                    {
-                    var students = await _dbContext.StudentInExamSchedule.ToListAsync();
-                    List<StudentInExamScheduleDTO> studentInExamScheduleDTOs = new List<StudentInExamScheduleDTO>();
-                    foreach (var student in students)
-                        {
-                        StudentInExamScheduleDTO studentInExamScheduleDTO = new StudentInExamScheduleDTO();
-                        studentInExamScheduleDTO.CopyProperties(student);
-                        studentInExamScheduleDTOs.Add(studentInExamScheduleDTO);
-                        await _dbContext.SaveChangesAsync();
-                        }
-                    return studentInExamScheduleDTOs;
-                    }
-                }
-            catch (Exception ex)
-                {
-                throw new Exception(ex.Message);
-                }
-            }
-        #endregion
-
-        #region Get All Student in Exam Schedule by ExamScheduleID
-        /// <summary>
-        ///  Get All Student In Exam Schedule by ExamScheduleID
-        /// </summary>
         /// <param name="examScheduleId"></param>
         /// <returns></returns>
-        /// <exception cref="Exception"></exception>
-        public async Task<List<StudentInExamScheduleDTO>> GetAllStudentInExamScheduleByExamScheduleId(int examScheduleId)
+        public async Task<List<StudentInExamSchedule>> GetAllStudentInExamScheduleByExamScheduleId(int examScheduleId)
             {
             try
                 {
-                using (var _dbContext = new MyDbContext())
+                using (var dbContext = new MyDbContext())
                     {
-                    var students = await _dbContext.StudentInExamSchedule.Where(s=> s.ExamScheduleId == examScheduleId).ToListAsync();
-                    List<StudentInExamScheduleDTO> studentInExamScheduleDTOs = new List<StudentInExamScheduleDTO>();
-                    foreach (var student in students)
-                        {
-                        StudentInExamScheduleDTO studentInExamScheduleDTO = new StudentInExamScheduleDTO();
-                        studentInExamScheduleDTO.CopyProperties(student);
-                        studentInExamScheduleDTOs.Add(studentInExamScheduleDTO);
-                        await _dbContext.SaveChangesAsync();
-                        }
-                    return studentInExamScheduleDTOs;
+                    return await dbContext.StudentInExamSchedule
+                        .Where(s => s.ExamScheduleId==examScheduleId)
+                        .ToListAsync();
                     }
                 }
             catch (Exception ex)
@@ -73,21 +37,19 @@ namespace SchedulerService.Repository.StudentInExamScheduleRepository
 
         #region Get Student in Exam Schedule by ExamScheduleID and StudentId
         /// <summary>
-        ///  Get Student in Exam Schedule by ExamScheduleID and StudentId
+        /// Get Student in Exam Schedule by ExamScheduleID and StudentId
         /// </summary>
         /// <param name="examScheduleId"></param>
+        /// <param name="studentId"></param>
         /// <returns></returns>
-        /// <exception cref="Exception"></exception>
-        public async Task<StudentInExamScheduleDTO> GetStudentInExamScheduleByExamScheduleIdAndStudentId(int examScheduleId, string studentId)
+        public async Task<StudentInExamSchedule> GetStudentInExamScheduleByExamScheduleIdAndStudentId(int examScheduleId, string studentId)
             {
             try
                 {
-                using (var _dbContext = new MyDbContext())
+                using (var dbContext = new MyDbContext())
                     {
-                    var student = await _dbContext.StudentInExamSchedule.FirstOrDefaultAsync(s => s.ExamScheduleId==examScheduleId && s.StudentId == studentId);
-                    StudentInExamScheduleDTO studentDto = new StudentInExamScheduleDTO();
-                    studentDto.CopyProperties(student);
-                    return studentDto;
+                    return await dbContext.StudentInExamSchedule
+                        .FirstOrDefaultAsync(s => s.ExamScheduleId==examScheduleId&&s.StudentId==studentId);
                     }
                 }
             catch (Exception ex)
@@ -101,17 +63,14 @@ namespace SchedulerService.Repository.StudentInExamScheduleRepository
         /// <summary>
         /// Add new Student into Exam Schedule Class
         /// </summary>
-        /// <param name="examScheduleDTO"></param>
+        /// <param name="student"></param>
         /// <returns></returns>
-        /// <exception cref="Exception"></exception>
-        public async Task<bool> AddNewStudentToExamSchedule(StudentInExamScheduleDTO examScheduleDTO)
+        public async Task<bool> AddNewStudentToExamSchedule(StudentInExamSchedule student)
             {
             try
                 {
                 using (var dbContext = new MyDbContext())
                     {
-                    StudentInExamSchedule student = new StudentInExamSchedule();
-                    student.CopyProperties(examScheduleDTO);
                     await dbContext.StudentInExamSchedule.AddAsync(student);
                     await dbContext.SaveChangesAsync();
                     return true;
@@ -128,17 +87,19 @@ namespace SchedulerService.Repository.StudentInExamScheduleRepository
         /// <summary>
         /// Remove student from exam schedule class
         /// </summary>
-        /// <param name="studentId"></param>
+        /// <param name="studentExamId"></param>
         /// <returns></returns>
-        /// <exception cref="Exception"></exception>
         public async Task<bool> RemoveStudentInExamScheduleClass(int studentExamId)
             {
             try
                 {
                 using (var dbContext = new MyDbContext())
                     {
-                    StudentInExamSchedule student=await dbContext.StudentInExamSchedule.FirstOrDefaultAsync(s => s.StudentExamId==studentExamId);
-                     dbContext.StudentInExamSchedule.Remove(student);
+                    var student = await dbContext.StudentInExamSchedule
+                        .FirstOrDefaultAsync(s => s.StudentExamId==studentExamId);
+                    if (student==null)
+                        return false;
+                    dbContext.StudentInExamSchedule.Remove(student);
                     await dbContext.SaveChangesAsync();
                     return true;
                     }
@@ -150,29 +111,22 @@ namespace SchedulerService.Repository.StudentInExamScheduleRepository
             }
         #endregion
 
-        #region Add Mutiple Student into Exam Schedule
+        #region Add Multiple Student into Exam Schedule
         /// <summary>
-        /// Add Mutiple Student into Exam Schedule
+        /// Add Multiple Student into Exam Schedule
         /// </summary>
-        /// <param name="studentInExamScheduleDTOs"></param>
+        /// <param name="students"></param>
         /// <returns></returns>
-        /// <exception cref="Exception"></exception>
-        public async Task<bool> AddMultipleStudentsToExamSchedule(List<StudentInExamScheduleDTO> studentInExamScheduleDTOs)
+        public async Task<bool> AddMultipleStudentsToExamSchedule(List<StudentInExamSchedule> students)
             {
             try
                 {
                 using (var dbContext = new MyDbContext())
                     {
-                    List<StudentInExamSchedule> students = studentInExamScheduleDTOs.Select(dto =>
-                    {
-                        var studentInClass = new StudentInExamSchedule();
-                        studentInClass.CopyProperties(dto);
-                        return studentInClass;
-                    }).ToList();
                     await dbContext.StudentInExamSchedule.AddRangeAsync(students);
                     await dbContext.SaveChangesAsync();
+                    return true;
                     }
-                return true;
                 }
             catch (Exception ex)
                 {
@@ -187,46 +141,14 @@ namespace SchedulerService.Repository.StudentInExamScheduleRepository
         /// </summary>
         /// <param name="examScheduleId"></param>
         /// <returns></returns>
-        /// <exception cref="Exception"></exception>
         public async Task<int> CountStudentInExamSchedule(int examScheduleId)
-            {
-            try
-                {
-                int count = 0;
-                using (var dbContext= new MyDbContext())
-                    {
-                     count = await dbContext.StudentInExamSchedule.CountAsync(s => s.ExamScheduleId==examScheduleId);
-                    }
-                return count;
-                }
-            catch (Exception ex)
-                {
-                throw new Exception(ex.Message);
-                }
-            }
-        #endregion
-
-        #region Get Student in Specific Class
-        /// <summary>
-        /// Check if a student already exists in a specific class exam schedule
-        /// </summary>
-        /// <param name="studentId"></param>
-        /// <param name="examScheduleId"></param>
-        /// <returns>StudentInExamScheduleDTO if exists, otherwise null</returns>
-        public async Task<StudentInExamScheduleDTO> GetStudentInClassByStudentIdAndClass(string studentId, int examScheduleId)
             {
             try
                 {
                 using (var dbContext = new MyDbContext())
                     {
-                    var studentInClass = await dbContext.StudentInExamSchedule.FirstOrDefaultAsync(s => s.StudentId==studentId&&s.ExamScheduleId==examScheduleId);
-                    if (studentInClass==null)
-                        {
-                        return null;
-                        }
-                    StudentInExamScheduleDTO studentInClassDTO = new StudentInExamScheduleDTO();
-                    studentInClassDTO.CopyProperties(studentInClass);
-                    return studentInClassDTO;
+                    return await dbContext.StudentInExamSchedule
+                        .CountAsync(s => s.ExamScheduleId==examScheduleId);
                     }
                 }
             catch (Exception ex)
@@ -235,6 +157,5 @@ namespace SchedulerService.Repository.StudentInExamScheduleRepository
                 }
             }
         #endregion
-
         }
     }
